cli: move common methods to util
